fix: refactor status labels for cd and sm
